requirements.txt
tqdm==4.59.0
tensorflow==2.5.1
numpy==1.19.5
wget==3.2
pandas==1.2.3
omegaconf==2.1.0
hydra_core==1.1.0
PyYAML==5.4.1
scikit_learn==0.24.2
dvc
dvc[s3]
streamlit
